apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "textpulse-api"
version = "0.1.0"
description = "FastAPI backend for TextPulse AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.1",
  "anthropic>=0.51.0",
  "boto3>=1.38.0",
  "celery>=5.4.0",
  "cryptography>=44.0.2",
  "email-validator>=2.2.0",
  "fastapi>=0.115.12",
  "httpx>=0.28.1",
  "ijson>=3.4.0",
  "lxml>=5.4.0",
  "orjson>=3.10.16",
  "pgvector>=0.3.6",
  "Pillow>=11.2.1",
  "psycopg[binary]>=3.2.6",
  "pydantic-settings>=2.8.1",
  "PyJWT>=2.10.1",
  "python-multipart>=0.0.20",
  "pytesseract>=0.3.13",
  "redis>=5.2.1",
  "sqlalchemy>=2.0.40",
  "slowapi>=0.1.9",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
